requirements.txt
garminconnect>=0.2.8
garth>=0.4.0
python-dotenv>=1.0
requests>=2.28.0
